\name{CollectLegacyMeasurements}
\alias{CollectLegacyMeasurements}
\title{Collect Measurement data frames from legacy CSV data}
\usage{
  CollectLegacyMeasurements(epochs)
}
\arguments{
  \item{epochs}{list of Epochs}
}
\value{
  Data Frame matching legacy CSV columns
}
\description{
  Collect Measurement data frames from legacy CSV data
}

